+++ include/footer.php
@@ -1,5 +1,5 @@
 <div class="row mt-3 mb-3 text-muted text-center">
       <div class="col-md text-left">
-        <p>Powered by <a href="https://github.com/laim/MyHomeBoardPHP">MyHomeBoardPHP</a> <?php print(MHBPHPVersion);?>.</p>
+        <p>Powered by <a href="https://github.com/laim/MyHomeBoardPHP">MyHomeBoardPHP</a><?php if(MHBPHP_HideVersion == false) { print(" " . MHBPHP_Version); }?>.</p>
       </div>
     </div>